+++ R/utils.R
@@ -12,9 +12,7 @@ quiet <- function(x) {
 }
 
 #------------------------------------------------
-#' Custom ggplot2 theme
-#'
-#' Custom `ggplot2` theme that builds off of `ggplot2::theme_classic()`.
+#' Custom ggplot2 themes
 #'
 #' @inheritParams ggplot2::theme_classic
 #'
@@ -25,10 +23,12 @@ quiet <- function(x) {
 #' p <- ggplot(mtcars, aes(x = wt, y = mpg, colour = factor(gear))) +
 #'   geom_point() +
 #'   facet_wrap(~am) +
-#'   geom_smooth(method = "lm", se = FALSE) +
-#'   theme_miplicorn()
-#' p
-theme_miplicorn <- function(base_size = 10, base_family = "",
+#'   geom_smooth(method = "lm", se = FALSE)
+#
+#' p + theme_miplicorn()
+#' p + theme_rainbow()
+theme_miplicorn <- function(base_size = 10,
+                            base_family = "",
                             base_line_size = base_size / 22,
                             base_rect_size = base_size / 22) {
   ggplot2::theme_classic(
@@ -44,10 +44,36 @@ theme_miplicorn <- function(base_size = 10, base_family = "",
     )
 }
 
+#' @rdname theme_miplicorn
+#' @export
+theme_rainbow <- function(base_size = 12,
+                          base_family = "",
+                          base_line_size = base_size / 22,
+                          base_rect_size = base_size / 22) {
+  ggplot2::theme_bw(
+    base_size = base_size,
+    base_family = base_family,
+    base_line_size = base_line_size,
+    base_rect_size = base_rect_size
+  ) %+replace%
+    ggplot2::theme(
+      plot.title = ggplot2::element_text(hjust = 0.5),
+      axis.line.x = ggplot2::element_line(color="black", size = 0.3),
+      axis.line.y = ggplot2::element_line(color="black", size = 0.3),
+      panel.grid.major = ggplot2::element_blank(),
+      panel.grid.minor = ggplot2::element_blank(),
+      legend.position = "bottom",
+      complete = TRUE
+    )
+}
+
 # Need to have a function that returns the default theme which will be used in
 # figures. This is according to the best practices for using ggplot2 in a
 # package. Link:
 # <https://ggplot2.tidyverse.org/articles/ggplot2-in-packages.html#creating-a-new-theme-1>
 default_theme <- function() {
   theme_miplicorn()
+}
+rainbow_theme <- function() {
+  theme_rainbow()
 }